Added better structure for graph fixtures
