--- tests/shapiq/graph/fixtures/__init__.py
@@ -1,2 +1,0 @@
-
-
