+++ tests/shapiq/graph/test_graphshapiq.py
@@ -16,72 +16,6 @@ from shapiq.graph.graphshapiq import GraphSHAPIQ
 class TestGraphSHAPIQ:
     """Test class for GraphSHAPIQ.__init__."""
 
-    @pytest.fixture
-    def gcn_graph_game(self, gcn_model, simple_graph):
-        return GraphGame(
-            model=gcn_model, x_graph=simple_graph, task="regression", baseline_strategy="average"
-        )
-
-    @pytest.fixture
-    def gin_graph_game(self, gin_model, simple_graph):
-        return GraphGame(
-            model=gin_model, x_graph=simple_graph, task="regression", baseline_strategy="average"
-        )
-
-    @pytest.fixture
-    def gat_graph_game(self, gat_model, simple_graph):
-        return GraphGame(
-            model=gat_model, x_graph=simple_graph, task="regression", baseline_strategy="average"
-        )
-
-    @pytest.fixture
-    def gcn_graph_game_small(self, gcn_model, small_graph):
-        return GraphGame(
-            model=gcn_model, x_graph=small_graph, task="regression", baseline_strategy="average"
-        )
-
-    @pytest.fixture
-    def gcn_graph_game_disconnected(self, gcn_model, disconnected_graph):
-        return GraphGame(
-            model=gcn_model,
-            x_graph=disconnected_graph,
-            task="regression",
-            baseline_strategy="average",
-        )
-
-    @pytest.fixture
-    def gcn_graph_game_single_node(self, gcn_model, single_node_graph):
-        return GraphGame(
-            model=gcn_model,
-            x_graph=single_node_graph,
-            task="regression",
-            baseline_strategy="average",
-        )
-
-    @pytest.fixture
-    def gcn_graphshapiq(self, gcn_graph_game):
-        return GraphSHAPIQ(game=gcn_graph_game)
-
-    @pytest.fixture
-    def gin_graphshapiq(self, gin_graph_game):
-        return GraphSHAPIQ(game=gin_graph_game)
-
-    @pytest.fixture
-    def gat_graphshapiq(self, gat_graph_game):
-        return GraphSHAPIQ(game=gat_graph_game)
-
-    @pytest.fixture
-    def gcn_graphshapiq_small(self, gcn_graph_game_small):
-        return GraphSHAPIQ(game=gcn_graph_game_small)
-
-    @pytest.fixture
-    def gcn_graphshapiq_disconnected(self, gcn_graph_game_disconnected):
-        return GraphSHAPIQ(game=gcn_graph_game_disconnected)
-
-    @pytest.fixture
-    def gcn_graphshapiq_single_node(self, gcn_graph_game_single_node):
-        return GraphSHAPIQ(game=gcn_graph_game_single_node)
-
     def test_init_with_gcn_model(self, gcn_graphshapiq, gcn_graph_game):
         """Test that GraphSHAPIQ initializes correctly with a GCN model."""
         assert gcn_graphshapiq.n_players == gcn_graph_game.n_players
